Move payments_table and table directories into shared
