+++ src/admin/containers/payment_reports.js
@@ -1,19 +1,19 @@
 /* @flow */
-import React, { createClass } from 'react'
+import React from 'react'
 import { connect } from 'react-redux'
 const { compose, props, map, append, lensIndex, set, range, apply, lift,
-  defaultTo, lensProp, reduce, over, keysIn, indexOf, identity }
+  defaultTo, lensProp, reduce, over, identity }
     = require('ramda')
 import { minus, plus } from 'app/money_arith'
 import standardise from 'app/standardise_date'
 import { fields, headers } from '../form_fields/paying_in.js'
 import Field from '../../shared/dumb_components/field.js'
-import MoneyRow from '../components/table/money_row.js'
+import MoneyRow from '../../shared/components/table/money_row.js'
 
 import { receive_non_cheque, receive_paying_in }
   from '../redux/modules/payment_reports.js'
 
-import Table from '../components/table'
+import Table from '../../shared/components/table'
 
 import { formatPounds } from 'app/monies'
