+++ src/admin/redux/reducer.js
@@ -3,7 +3,7 @@ import { combineReducers } from 'redux'
 import form from './form_reducer.js'
 import payment_defaults from './modules/payment_defaults.js'
 import mode from '../../shared/redux/modules/mode.js'
-import payments from './modules/payments.js'
+import payments from '../../shared/redux/modules/payments.js'
 import charge_form from './modules/charge_form.js'
 import route from '../../shared/redux/modules/route.js'
 import payment_reports from './modules/payment_reports.js'
