+++ src/shared/components/payments_table/deletion_entry.js
