+++ src/shared/components/payments_table/index.js
