+++ src/shared/components/table/entry.js
