+++ src/shared/components/table/index.js
