+++ src/shared/components/table/money_row.js
