+++ src/shared/components/table/smart_row.js
