+++ src/user/redux/reducer.js
@@ -3,7 +3,7 @@ import { combineReducers } from 'redux'
 import my_details_form from './form_reducer.js'
 
 //reducers
-import payments from '../../admin/redux/modules/payments.js'
+import payments from '../../shared/redux/modules/payments.js'
 import mode from '../../shared/redux/modules/mode.js'
 import active_tab from './modules/active_tab.js'
